% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2DtoGridDF.R
\name{convert2DtoGridDF}
\alias{convert2DtoGridDF}
\title{Create a \pkg{tibble} tibble representing a ROMS 2d spatial variable}
\usage{
convert2DtoGridDF(m, name = "value")
}
\arguments{
\item{m}{- matrix (xi x eta) from call to}

\item{name}{- name for the `value` column (defaults to "value")}
}
\value{
A \pkg{tibble} tibble object representing the grid
}
\description{
Function to create a \pkg{tibble} tibble representing a ROMS 2d spatial variable.
}
\details{
Creates a tibble from a matrix representing a ROMS 2d spatial variable.
The matrix should (probably) represent the full xi,eta dimensions of the ROMS model. Any
decimation should be done afterward on the resulting tibble. This will ensure the xi, eta
indices in the dataframe correspond to the correct grid cells in the original output.

The returned \pkg{sf} object has columns:
\itemize{
  \item{ID - string "xi_eta"}
  \item{xi - index value}
  \item{eta - eta index value}
  \item{`name` - matrix value at `[xi,eta]`}
}
}
