% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBasemap.R
\name{createBasemap}
\alias{createBasemap}
\title{Create a basemap}
\usage{
createBasemap(
  boundingbox = tmaptools::bb(xlim = c(-162, -134), ylim = c(54, 62)),
  colors.background = "grey85",
  colors.land = "grey45",
  colors.bathym = "grey65",
  aes.palette = list(cat = c("blue", "red"), div = c("blue", "red"), seq = c("blue",
    "red"))
)
}
\arguments{
\item{boundingbox}{- sf-style bounding box for map (default is for CGOA)}

\item{colors.background}{- background color}

\item{colors.land}{- color for land}

\item{colors.bathym}{- color for bathymetry}

\item{aes.palette}{- list with color palettes (or vectors) for cat, div, and seq "by" variables}
}
\value{
tmap-style basemap
}
\description{
Function to create a basemap.
}
\details{
Requires packages \code{wtsGIS}, \code{tmap}, and possibly tmaptools.
See \code{wtsGIS::getStandardBoundingBox} for alternative, predefined bounding boxes.
}
