% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGridFromNetCDF.R
\name{createGridFromNetCDF}
\alias{createGridFromNetCDF}
\title{Create a \code{sf} dataframe representing a ROMS grid from a netcdf file}
\usage{
createGridFromNetCDF(
  netcdf,
  grid_type = c("rho", "psi", "u", "v"),
  coord_type = c("latlon", "xy"),
  strCRS = ifelse(tolower(coord_type[1]) == "latlon", wtsGIS::getCRS("NAD83"),
    wtsGIS::getCRS("AlaskaAlbers")),
  test = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{netcdf}{- path to netcdf file with ROMS grid, or a ncdf4 object connected to a netcdf file}

\item{strCRS}{- coordinate reference for grid}
}
\value{
A \code{sf} dataframe object representing the grid
}
\description{
Function to create a \code{sf} dataframe representing a ROMS grid from a netcdf file.
}
\details{
Creates an sf dataframe from
a netcdf file representing a ROMS grid.
}
