% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGridFromShapefile.R
\name{createGridFromShapefile}
\alias{createGridFromShapefile}
\title{Create a \code{sf} dataframe representing a ROMS grid from a shapefile}
\usage{
createGridFromShapefile(shapefile, strCRS = wtsGIS::get_crs("AlaskaAlbers")[1])
}
\arguments{
\item{shapefile}{- path to shapefile derived from ROMS grid.}

\item{strCRS}{- coordinate reference for a shapefile}
}
\value{
A \code{sf} dataframe object representing the grid
}
\description{
Function to create a \code{sf} dataframe representing a ROMS grid from a shapefile.
}
\details{
Uses \code{wtsGIS::createLayerFromShapefile} to create an sf dataframe from
a shapefile derived from a ROMS grid.
}
