% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGrid.R
\name{getGrid}
\alias{getGrid}
\title{Get a ROMS model grid as an sf dataframe}
\usage{
getGrid(name, verbose = FALSE)
}
\arguments{
\item{name}{- name of grid to return}

\item{verbose}{- flag (T/F) to print additional information}
}
\description{
Function to get a ROMS model grid as an sf dataframe.
}
\details{
Returns an sf dataframe corresponding to the requested model grid.\cr
Available grids are:
\itemize{
 \item{"CGOA" - Coastal Gulf of Alaska 3km grid}
 \item{"Bering10K" - Bering Sea 10km grid}
}
}
