% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfo_ROMSFiles.R
\name{getInfo_ROMSFiles}
\alias{getInfo_ROMSFiles}
\title{Get ROMS file names corresponding to integration dates}
\usage{
getInfo_ROMSFiles(
  topDir,
  start = as.POSIXct("1982-01-01 00:00:00", tz = "GMT"),
  delt = 24 * 3600
)
}
\arguments{
\item{topDir}{- path to folder with ROMS output files}

\item{start}{- first time to include (as POSIXct object)}

\item{delt}{- integration timestep in seconds (default = 24*3600)}
}
\value{
a dataframe wth ROMS file information by date
}
\description{
Function to get ROMS file names corresponding to integration dates
}
