% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandardBBox.R
\name{getStandardBBox}
\alias{getStandardBBox}
\title{Create a bounding box to display a standard map area for ROMS model output}
\usage{
getStandardBBox(
  area = c("EBS", "Bering10k", "CGOA", "CGOA shelf"),
  type = "sf"
)
}
\arguments{
\item{area}{- area to return the bounding box for ("EBS","Bering10k","CGOA","CGOA shelf")}

\item{type}{- bounding box type to return ("sf" or "sp" for \pkg{sf} or \pkg{sp} types, respectively)}
}
\value{
- bounding box in \pkg{sf} (vector) or \pkg{sp} (list) format.
}
\description{
This function returns a bounding box to display a standard map area for ROMS model output.
}
\details{
These bounding boxes are slightly different
from those provided by \code{\link[wtsGIS]{getStandardBBBox}} and are optimized for
ROMS model output.

Available areas are: "EBS", "Bering10k", "CGOA", "CGOA shelf"
}
