% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandardBoundingBox.R
\name{getStandardBoundingBox}
\alias{getStandardBoundingBox}
\title{Create a bounding box to display a standard map area  for ROMS model output}
\usage{
getStandardBoundingBox(area = c("EBS", "Bering10k", "CGOA", "CGOA shelf"))
}
\arguments{
\item{area}{- area to return the bounding box for ("EBS","Bering10k","CGOA","CGOA shelf")}
}
\value{
- bounding box in list format
}
\description{
This function returns a bounding box to display a standard map area for ROMS model output.
}
\details{
The bounding box is returned in list format, with coordinates defining the
bottomleft and topright boundaries for the area. These bounding boxes are slightly different
from those provided by \code{\link[wtsGIS]{getStandardBoundingBox}} and are optimized for
ROMS model output.

Available areas are: "EBS", "Bering10k", "CGOA", "CGOA shelf"
}
