% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_CreateROMSGridLayers.R
\name{gg_CreateROMSGridLayers}
\alias{gg_CreateROMSGridLayers}
\title{Create ROMS grid layers for maps based on the \pkg{ggplot2} package}
\usage{
gg_CreateROMSGridLayers(
  grid,
  bathym.scale = c(seq(0, 200, 25), seq(250, 500, 50), seq(1000, 6000, 1000)),
  final.crs = wtsGIS::get_crs("WGS84"),
  bbox = wtsROMS::getStandardBBox("EBS"),
  colors.bg = "white",
  colors.land = "grey85",
  scale.bathym = ggplot2::scale_fill_viridis_d(option = "plasma", direction = -1),
  alpha.bathym = 1
)
}
\arguments{
\item{grid}{- a ROMS grid from call to \code{\link{getGrid}}}

\item{final.crs}{- representation of final (display) coordinate reference system (see details)}

\item{bbox}{- a bounding box (see details)}

\item{colors.bg}{- background color}

\item{colors.land}{- color for land}

\item{scale.bathym}{- \pkg{ggplot2} scale_fill_ object to use for the bathymetry fill scale}

\item{alpha.bathym}{- transparency for the bathymetry}
}
\value{
- a list with elements based on the \pkg{ggplot2} package:
\itemize{
\item{land - geom_sf layer for land-based grid polygons}
\item{bathym - geom_sf layer for bathymetry-based grid polygons}
\item{fill_scale - scale_fill_ object for bathymetry}
\item{map_scale - ggplot2 coord_ object}
\item{theme - partial ggplot2 theme with axis titles removed}
}
}
\description{
This function creates a ROMS grid layers for maps based on the \pkg{ggplot2} package.
}
\details{
The returned list contains land and bathymetry layers (polygons), a fill scale for the bathymetry layer,
a map scale, and a partial theme.

The final coordinate reference system (\code{final.crs}) can be any object that
can be converted to a \code{sf::crs} object using \code{\link{get_crs}}.

The bounding box (\code{bbox}) can be any object that can be converted
to a \code{sf::bbox} using \code{\link{getBBox}}.
}
