% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapLayer_CreateFromModelOutput.R
\name{mapLayer_CreateFromModelOutput}
\alias{mapLayer_CreateFromModelOutput}
\title{Create a \code{tmap} map layer from an \code{sf} dataset representing a model output}
\usage{
mapLayer_CreateFromModelOutput(obj, palette = NULL)
}
\arguments{
\item{obj}{- \code{sf} dataset}

\item{palette}{-}
}
\value{
A \code{tmap} shape/fill combination map layer with polygons colored using the palette
to represent values of the ROMS variable in obj.
}
\description{
Function to create a \code{tmap} map layer from an \code{sf} dataset representing a model output.
}
\details{
Creates a \code{tmap} shape/fill combination map layer with polygons colored using the palette
to represent values of the ROMS variable in obj. If not provided, the palette will default to the
palette associated with the basemap used to plot the layer.
}
