% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCDF_CopyFile.R
\name{netCDF_CopyFile}
\alias{netCDF_CopyFile}
\title{Copy a netCDF file, possibly dropping variables}
\usage{
netCDF_CopyFile(
  infn,
  outfn,
  varNames = NULL,
  varPatterns = NULL,
  action = c("keep", "drop"),
  test = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{infn}{- input file (file to copy)}

\item{outfn}{- output file}

\item{varNames}{- vector of variable names}

\item{varPatterns}{- pattern to match for variable names}

\item{action}{- what to do with specified variables ("keep" or "drop") on copy}

\item{test}{- flag (T/F) to test the copy process, but don't actually do it}

\item{verbose}{- flag to print extra information}
}
\value{
Dataframe with filename, ocean_time (in seconds), and ocean_date (POSIXct dates) as columns.
}
\description{
Function to copy a netCDF file, possibly dropping variables.
}
\details{
Numeric values for ocean_times are in seconds
relative to the calendar reference date.
}
