% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCDF_CreateGrid.R
\name{netCDF_CreateGrid}
\alias{netCDF_CreateGrid}
\title{Create a \code{sf} dataframe representing a ROMS grid from a netcdf file}
\usage{
netCDF_CreateGrid(
  netcdf,
  grid_type = c("rho", "psi", "u", "v"),
  coord_type = c("latlon", "xy"),
  strCRS = ifelse(tolower(coord_type[1]) == "latlon", wtsGIS::getCRS("NAD83"),
    wtsGIS::getCRS("AlaskaAlbers")),
  verbose = TRUE
)
}
\arguments{
\item{netcdf}{- path to netcdf file with ROMS grid, or a ncdf4 object connected to a netcdf file}

\item{grid_type}{-}

\item{coord_type}{-}

\item{strCRS}{- coordinate reference for grid}

\item{verbose}{- flag to print additional information}
}
\value{
A \code{sf} dataframe object representing the grid
}
\description{
Function to create a \code{sf} dataframe representing a ROMS grid from a netcdf file.
}
\details{
Creates an sf dataframe from a netcdf file representing a ROMS grid.
Masked cells (land) have a value of NA.
}
