% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCDF_CreateModelOutput.R
\name{netCDF_CreateModelOutput}
\alias{netCDF_CreateModelOutput}
\title{Create a \code{sf} dataframe representing a ROMS model output layer from a netcdf file}
\usage{
netCDF_CreateModelOutput(netcdf, roms_grid, variable, timeslice = 1, layer = 1)
}
\arguments{
\item{netcdf}{- path to netcdf file with ROMS model output, or a ncdf4 object connected to a netcdf file}

\item{roms_grid}{- roms grid as an sf dataframe object}

\item{variable}{- variable name to extract}

\item{timeslice}{- time slice from file to extract}

\item{layer}{- vertical layer to extract (1: bottom)}
}
\value{
A \code{sf} dataframe object representing the layer
}
\description{
Function to create a \code{sf} dataframe representing a ROMS model output layer from a netcdf file.
}
\details{
Creates an sf dataframe from
a netcdf file representing a ROMS model output layer.
}
