% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCDF_CreateROMSGrid.R
\name{netCDF_CreateROMSGrid}
\alias{netCDF_CreateROMSGrid}
\title{Create a \code{sf} dataframe representing a ROMS grid from a netcdf file}
\usage{
netCDF_CreateROMSGrid(
  netcdf,
  grid_type = c("rho", "psi", "u", "v"),
  coord_type = c("latlon", "xy"),
  crs = ifelse(tolower(coord_type[1]) == "latlon", wtsGIS::get_crs("WGS84")[[1]],
    sf::NA_crs_),
  verbose = TRUE
)
}
\arguments{
\item{netcdf}{- path to netcdf file with ROMS grid, or a ncdf4 object connected to a netcdf file}

\item{grid_type}{- "rho","psi","u","v" [only works now for "rho" grid type]}

\item{coord_type}{- "latlon","xy" [only works now for "latlon" type]}

\item{crs}{- coordinate reference for grid (latlon default is WGS84, xy default is sf::NA_crs_)}

\item{verbose}{- flag to print additional information}
}
\value{
A \code{sf} dataframe object representing the grid
}
\description{
Function to create a \code{sf} dataframe representing a ROMS grid from a netcdf file.
}
\details{
Creates an sf dataframe from a netcdf file representing a ROMS grid. 

The input \code{crs} can be anything that \code{\link[wtsGIS]{get_crs}} can convert to
an \pkg{sf}-style coordinate reference system object.
}
