% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCDF_GetInfoVarDims.R
\name{netCDF_GetInfoVarDims}
\alias{netCDF_GetInfoVarDims}
\title{Get information on dimensions for a variable in a netcdf file}
\usage{
netCDF_GetInfoVarDims(netcdf, variable, verbose = FALSE)
}
\arguments{
\item{netcdf}{- path to netcdf file (or previously opened connection)}

\item{variable}{- variable name to extract}

\item{verbose}{- flag (T/F) to print info}
}
\value{
vector of dimension sizes, with dimension names as element names
}
\description{
Function to get information on dimensions for a variable in a netcdf file.
}
\details{
None.
}
