% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCDF_GetOceanTimes.R
\name{netCDF_GetOceanTimes}
\alias{netCDF_GetOceanTimes}
\title{Get ocean_times from a ROMS model netcdf file}
\usage{
netCDF_GetOceanTimes(netcdf)
}
\arguments{
\item{netcdf}{- path to ROMS model netcdf file (or previously opened connection)}
}
\value{
Dataframe with columns ocean_times (in seconds since model origin) and dates (POSIXct dates).
}
\description{
Function to get ocean_times from a ROMS model netcdf file.
}
\details{
Dates associated with ocean_times are relative to the model reference date determined by
parsing the \code{units} associated with the \code{ocean_time} dimension in the netcdf dataset.
}
