% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCDF_Open.R
\name{netCDF_Open}
\alias{netCDF_Open}
\title{Open a netcdf file}
\usage{
netCDF_Open(netcdf)
}
\arguments{
\item{netcdf}{- path to netcdf file (or previously opened connection)}
}
\value{
object of class 'ncdf4', an opened connection to a netcdf file.
}
\description{
Function to open a netcdf file and return an opened connection.
}
\details{
The returned object should be closed by the user using\code{ncdf4::nc_close(...)}.
\code{netcdf} is simply passed through if it is already an open connection.
}
