% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanTime_Convert.R
\name{oceanTime_ConvertToNumeric}
\alias{oceanTime_ConvertToNumeric}
\title{Convert ocean_times from character or POSIXct to numeric}
\usage{
oceanTime_ConvertToNumeric(ot, ref = "1900-01-01 00:00:00")
}
\arguments{
\item{ot}{- ocean_time, as character or POSIXct vector}

\item{ref}{- calendar reference (default = "1900-01-01 00:00:00")}
}
\value{
Vector with ocean_times as numeric seconds since the reference date.
}
\description{
Function to convert ocean_times from character or POSIXct to numeric.
}
\details{
Input numeric values associated with ocean_times should be a vector of character strings
formatted as 'YYY-MM-DD HH:MM:SS' or POSIXct.
}
