% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanTime_Convert.R
\name{oceanTime_ConvertToPOSIXct}
\alias{oceanTime_ConvertToPOSIXct}
\title{Convert ocean_times from numeric to POSIXct}
\usage{
oceanTime_ConvertToPOSIXct(ot, ref = "1900-01-01 00:00:00")
}
\arguments{
\item{ot}{- ocean_time, as numeric vector}

\item{ref}{- calendar reference (default = "1900-01-01 00:00:00")}
}
\value{
Vector with ocean_times as POSIXct dates.
}
\description{
Function to convert ocean_times from numeric to character.
}
\details{
Input numeric values associated with ocean_times should be in seconds
relative to the model reference date.
}
