% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanTime_Find.R
\name{oceanTime_Find}
\alias{oceanTime_Find}
\title{Find the file in which an input ocean_time is bracketed}
\usage{
oceanTime_Find(
  ocean_time,
  path,
  pattern,
  ref = as.POSIXct("1900-01-01 00:00:00", tz = "GMT"),
  verbose = FALSE
)
}
\arguments{
\item{ocean_time}{- time (in seconds relative to the calendar reference or as POSIXct date)}

\item{path}{- path to folder with model output files}

\item{pattern}{- pattern for model output filenames (e.g. "avg_*.nc")}

\item{ref}{- calendar reference (default = "1900-01-01 00:00:00")}

\item{verbose}{- flag to print extra information}
}
\value{
Name of model output file in which the input ocean_time is bracketed.
}
\description{
Function to find the file in which an input ocean_time is bracketed.
}
\details{
None.
}
