% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceanTime_GetTimes.R
\name{oceanTime_GetTimes}
\alias{oceanTime_GetTimes}
\title{Get a dataframe of ocean_times associated with files in a folder}
\usage{
oceanTime_GetTimes(
  path,
  pattern,
  ref = as.POSIXct("1900-01-01 00:00:00", tz = "GMT"),
  verbose = FALSE
)
}
\arguments{
\item{path}{- path to folder with model output files}

\item{pattern}{- pattern for model output filenames (e.g. "avg_*.nc")}

\item{ref}{- calendar reference (default = "1900-01-01 00:00:00")}

\item{verbose}{- flag to print extra information}
}
\value{
Dataframe with filename, ocean_time (in seconds), and ocean_date (POSIXct dates) as columns.
}
\description{
Function to get a dataframe of ocean_times associated with files in a folder.
}
\details{
Numeric values for ocean_times are in seconds
relative to the calendar reference date.
}
