% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCDF_CreateGrid.R
\name{netCDF_CreateGrid}
\alias{netCDF_CreateGrid}
\title{Create a \code{sf} dataframe representing a ROMS grid from a netcdf file}
\usage{
netCDF_CreateGrid(
  netcdf,
  grid_type = c("rho", "psi", "u", "v"),
  coord_type = c("latlon", "xy"),
  strCRS = ifelse(tolower(coord_type[1]) == "latlon", wtsGIS::get_crs("NAD83")[[1]],
    wtsGIS::get_crs("AlaskaAlbers")[[1]]),
  verbose = TRUE
)
}
\arguments{
\item{netcdf}{- path to netcdf file with ROMS grid, or a ncdf4 object connected to a netcdf file}

\item{grid_type}{- "rho","psi","u","v" [only works now for "rho" grid type]}

\item{coord_type}{- "latlon","xy" [only works now for "latlon" type]}

\item{strCRS}{- coordinate reference for grid}

\item{verbose}{- flag to print additional information}
}
\value{
A \code{sf} dataframe object representing the grid
}
\description{
Function to create a \code{sf} dataframe representing a ROMS grid from a netcdf file.
}
\details{
Creates an sf dataframe from a netcdf file representing a ROMS grid.
Masked cells (e.g., land) have a value of NA.
}
