% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCDF_GetInfo.R
\name{netCDF_GetInfo}
\alias{netCDF_GetInfo}
\title{Get information on dimensions and variables in a netcdf file}
\usage{
netCDF_GetInfo(netcdf, verbose = FALSE)
}
\arguments{
\item{netcdf}{- path to netcdf file (or previously opened connection)}

\item{verbose}{- flag to print debugging info}
}
\value{
list of dataframes with information on dimensions ("dims") and variables ("vars").
}
\description{
Function to get information on dimensions and variables in a netcdf file.
}
\details{
None.
}
